Move prom stuff to prometheus folder
