+++ go-sdk/prometheus/panels/time-series/time-series.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package bar
+package time_series
 
 import (
 	commonSdk "github.com/perses/perses/go-sdk/common"
