+++ go-sdk/prometheus/variables/prometheus-label-names/prometheus-label-names.go
@@ -11,24 +11,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package prometheus_label_values
+package prometheus_label_names
 
-import "github.com/perses/perses/pkg/model/api/v1/common"
-
-type datasourceSelector struct {
-	Kind string `json:"kind" yaml:"kind"`
-	Name string `json:"name,omitempty" yaml:"name,omitempty"`
-}
+import (
+	"github.com/perses/perses/go-sdk/prometheus/variables"
+	"github.com/perses/perses/pkg/model/api/v1/common"
+)
 
 type PluginSpec struct {
-	Datasource *datasourceSelector `json:"datasource,omitempty" yaml:"datasource,omitempty"`
-	LabelName  string              `json:"labelName" yaml:"labelName"`
-	Matchers   []string            `json:"matchers,omitempty" yaml:"matchers,omitempty"`
+	Datasource *variables.DatasourceSelector `json:"datasource,omitempty" yaml:"datasource,omitempty"`
+	Matchers   []string                      `json:"matchers,omitempty" yaml:"matchers,omitempty"`
 }
 
-func NewLabelValuesVariablePlugin(labelName string) *ListVariablePluginBuilder {
+func NewLabelNamesVariablePlugin() *ListVariablePluginBuilder {
 	return &ListVariablePluginBuilder{
-		PluginSpec: PluginSpec{LabelName: labelName},
+		PluginSpec: PluginSpec{},
 	}
 }
 
@@ -38,14 +35,9 @@ type ListVariablePluginBuilder struct {
 
 func (b *ListVariablePluginBuilder) Build() common.Plugin {
 	return common.Plugin{
-		Kind: "PrometheusLabelValuesVariable",
+		Kind: "PrometheusLabelNamesVariable",
 		Spec: b.PluginSpec,
 	}
-}
-
-func (b *ListVariablePluginBuilder) WithLabelName(label string) *ListVariablePluginBuilder {
-	b.LabelName = label
-	return b
 }
 
 func (b *ListVariablePluginBuilder) WithMatchers(matchers []string) *ListVariablePluginBuilder {
@@ -59,7 +51,7 @@ func (b *ListVariablePluginBuilder) AddMatcher(matcher string) *ListVariablePlug
 }
 
 func (b *ListVariablePluginBuilder) WithDatasource(name string) *ListVariablePluginBuilder {
-	b.Datasource = &datasourceSelector{
+	b.Datasource = &variables.DatasourceSelector{
 		Kind: "PrometheusDatasource",
 		Name: name,
 	}
