+++ go-sdk/prometheus/variables/selector.go
@@ -1,0 +1,6 @@
+package variables
+
+type DatasourceSelector struct {
+	Kind string `json:"kind" yaml:"kind"`
+	Name string `json:"name,omitempty" yaml:"name,omitempty"`
+}
